Factor ObserveFields test code into helpers header
